@@ -1,53 +1,118 @@
 {
-  "name": "js-use-core",
-  "version": "1.0.0",
-  "description": "JavaScript 综合类型工具库，包含全屏、复制粘贴等功能",
-  "main": "dist/index.js",
-  "module": "dist/index.esm.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "lint": "eslint src --ext .ts",
-    "lint:fix": "eslint src --ext .ts --fix",
-    "type-check": "tsc --noEmit",
-    "clean": "rimraf dist",
-    "prebuild": "npm run clean"
-  },
-  "keywords": [
-    "javascript",
-    "typescript",
-    "utility",
-    "fullscreen",
-    "clipboard",
-    "copy",
-    "paste"
-  ],
-  "author": "Your Name",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/jest": "^29.5.0",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "eslint": "^8.45.0",
-    "jest": "^29.5.0",
-    "rimraf": "^5.0.0",
-    "rollup": "^3.28.0",
-    "rollup-plugin-typescript2": "^0.35.0",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.1.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/your-username/js-use-core.git"
-  },
-  "bugs": {
-    "url": "https://github.com/your-username/js-use-core/issues"
-  },
-  "homepage": "https://github.com/your-username/js-use-core#readme"
-} 
+	"name": "js-use-core",
+	"version": "1.1.0",
+	"description": "JavaScript 综合类型工具库，包含全屏、复制粘贴等功能",
+	"keywords": [
+		"javascript",
+		"typescript",
+		"utility",
+		"fullscreen",
+		"clipboard",
+		"copy",
+		"paste",
+		"file",
+		"image",
+		"base64",
+		"blob",
+		"conversion",
+		"utility",
+		"dataURL",
+		"image-compression",
+		"image-conversion",
+		"file-conversion",
+		"font",
+		"font-loading",
+		"font-check",
+		"font-face",
+		"css-font-loading-api",
+		"web-fonts",
+		"font-management",
+		"typescript",
+		"javascript"
+	],
+	"homepage": "https://github.com/chao921125/js-use-core",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/chao921125/js-use-core.git"
+	},
+	"bugs": {
+		"url": "https://github.com/chao921125/js-use-core/issues"
+	},
+	"author": {
+		"name": "chao921125",
+		"email": "chao921125@sina.com",
+		"url": "https://github.com/chao921125"
+	},
+	"license": "BSD-3-Clause",
+	"type": "module",
+	"main": "dist/index.js",
+	"module": "dist/index.esm.js",
+	"types": "dist/types/index.d.ts",
+	"unpkg": "dist/index.umd.js",
+	"jsdelivr": "dist/index.umd.js",
+	"files": [
+		"dist"
+	],
+	"scripts": {
+		"build": "rollup -c",
+		"dev": "rollup -c -w",
+		"test": "jest",
+		"test:watch": "jest --watch",
+		"lint": "eslint src --ext .ts",
+		"lint:fix": "eslint src --ext .ts --fix",
+		"type-check": "tsc --noEmit",
+		"clean": "rimraf dist",
+		"prebuild": "npm run clean"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.28.0",
+		"@babel/preset-env": "^7.28.0",
+		"@babel/preset-typescript": "^7.27.1",
+		"@rollup/plugin-babel": "^6.0.4",
+		"@rollup/plugin-commonjs": "^28.0.6",
+		"@rollup/plugin-json": "^6.1.0",
+		"@rollup/plugin-node-resolve": "^16.0.1",
+		"@rollup/plugin-terser": "^0.4.4",
+		"@rollup/plugin-typescript": "^12.1.4",
+		"@testing-library/jest-dom": "^6.6.3",
+		"@types/jest": "^29.5.14",
+		"@types/node": "^24.0.13",
+		"@typescript-eslint/eslint-plugin": "^6.21.0",
+		"@typescript-eslint/parser": "^6.21.0",
+		"eslint": "^8.57.1",
+		"jest": "^29.7.0",
+		"jest-environment-jsdom": "^30.0.4",
+		"rimraf": "^5.0.10",
+		"rollup": "^3.29.5",
+		"rollup-plugin-dts": "^6.2.1",
+		"rollup-plugin-terser": "^7.0.2",
+		"rollup-plugin-typescript2": "^0.35.0",
+		"ts-jest": "^29.4.0",
+		"tslib": "^2.8.1",
+		"typescript": "^5.8.3"
+	},
+	"overrides": {},
+	"engines": {
+		"node": ">=14.0.0",
+		"npm": ">=6.14.18",
+		"yarn": ">=3.0.0"
+	},
+	"browserslist": [
+		"defaults",
+		"> 1%",
+		"not dead",
+		"not op_mini all",
+		"last 2 versions",
+		"Chrome > 51",
+		"Edge > 15",
+		"Firefox > 54",
+		"Safari > 10",
+		"not IE <= 11",
+		"Android >= 6.0",
+		"iOS > 10",
+		"supports es6-module"
+	],
+	"publishConfig": {
+		"access": "public"
+	}
+}
